pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "aind-airflow-jobs"
description = "Global classes for AIND Airflow service"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Allen Institute for Neural Dynamics"}
]
classifiers = [
    "Programming Language :: Python :: 3"
]
readme = "README.md"
dynamic = ["version"]

dependencies = [
    "google-re2==1.0",
    "apache-airflow==2.8.1",
    "apache-airflow-providers-ssh",
    "aind-slurm-rest==1.0.3",
    "python-logging-loki==0.3.1",
    "pydantic-settings",
    "pydantic"
]

[project.optional-dependencies]
dev = [
    'black',
    'coverage',
    'flake8',
    'interrogate',
    'isort',
    'Sphinx',
    'furo'
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.dynamic]
version = {attr = "aind_airflow_jobs.__version__"}

[tool.black]
line-length = 79
target_version = ['py38']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | build
    | dist
  )/
  | .gitignore
)
'''

[tool.coverage.run]
omit = ["*__init__*"]
source = ["aind_airflow_jobs", "tests"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
    "from",
    "import",
    "pragma: no cover"
]
fail_under = 100

[tool.isort]
line_length = 79
profile = "black"

[tool.interrogate]
exclude = ["setup.py", "docs", "build"]
fail-under = 100
